// src/modules/wish-history/interfaces/excel.interface.ts
export interface ITypeAmount {
  type: string;
  amount: number;
}

export interface IExcelRow {
  A: string;
  B: string;
  C: string;
  D: number;
  E: number;
  F: number;
  G: number;
  H: string;
  I: string;
}

export interface IExcelRowTitle {
  Type: string;
  Name: string;
  Time: string;
  Rarity: number;
  Pity: number;
  Roll: number;
  Group: number;
  Title: string;
  Part: string;
}